Orca: Merge TorchRunner with TrainingOperator(validate) (#6880)
* trainingOp val

* val in train_epochs
