+++ python/orca/src/bigdl/orca/learn/pytorch/training_operator.py
@@ -137,55 +137,6 @@ class TrainingOperator:
         """
         pass
 
-    def validate(self, val_iterator, info, metrics, num_steps=None):
-        """Runs one standard validation pass over the val_iterator.
-
-        This will call ``model.eval()`` and ``torch.no_grad`` when iterating
-        over the validation dataloader.
-
-        If overriding this method, you can access model, criterion via
-        ``self.model`` and ``self.criterion``. You also do not need to call
-        ``validate_batch`` if overriding this method.
-
-        Args:
-            val_iterator (iter): Iterable constructed from the
-                validation dataloader.
-            info: (dict): Dictionary for information to be used for custom
-                validation operations.
-
-        Returns:
-            A dict of metrics from the evaluation.
-                By default, returns "val_accuracy" and "val_loss"
-                which is computed by aggregating "loss" and "correct" values
-                from ``validate_batch`` and dividing it by the sum of
-                ``num_samples`` from all calls to ``self.validate_batch``.
-        """
-        # switch to evaluate mode
-        self.model.eval()
-        metrics = Metric.convert_metrics_dict(metrics, backend="pytorch")
-        losses = []
-        total_samples = 0
-        with torch.no_grad():
-            for batch_idx, batch in enumerate(val_iterator):
-                if num_steps and batch_idx == num_steps:
-                    break
-                batch_info = {"batch_idx": batch_idx}
-                batch_info.update(info)
-                output, target, loss = self.forward_batch(batch, batch_info)
-                num_samples = get_batchsize(target)
-                total_samples += num_samples
-                losses.append(loss.item() * num_samples)
-                for metric in metrics.values():
-                    metric(output, target)
-
-        result = {name: metric.compute() for name, metric in metrics.items()}
-
-        result["val_loss"] = sum(losses) / total_samples
-
-        result["num_samples"] = total_samples
-
-        return result
-
     def predict(self, pred_iterator):
         # switch to evaluate mode
         self.model.eval()
@@ -213,50 +164,6 @@ class TrainingOperator:
         # todo support multi-output model
         np_output = output.detach().numpy()
         return np_output
-
-    def forward_batch(self, batch, batch_info):
-        """Calculates the loss and accuracy over a given batch.
-
-        You can override this method to provide arbitrary metrics.
-
-        Same as ``train_batch``, this method implementation assumes that
-        batches are in (\*features, labels) format by default. So we also
-        support multiple inputs model.
-
-        Args:
-            batch: One item of the validation iterator.
-            batch_info (dict): Contains information per batch from
-                ``validate()``.
-
-        Returns:
-            A dict of metrics.
-                By default, returns "val_loss", "val_accuracy", and
-                "num_samples". When overriding, consider returning
-                "num_samples" in the metrics because
-                by default, ``validate`` uses "num_samples" to
-                calculate averages.
-        """
-        # unpack features into list to support multiple inputs model
-        features, target = batch
-
-        # compute output
-        with self.timers.record("eval_fwd"):
-            if torch.is_tensor(features):
-                output = self.model(features)
-            elif isinstance(features, (tuple, list)):
-                output = self.model(*features)
-            else:
-                invalidInputError(False,
-                                  "Features should be tensor, list/tuple or dict, "
-                                  "but got {}".format(type(features)))
-
-            if isinstance(output, tuple) or isinstance(output, list):
-                # Then target is also assumed to be a tuple or list.
-                loss = self.criterion(*output, *target)
-            else:
-                loss = self.criterion(output, target)
-
-        return output, target, loss
 
     def state_dict(self):
         """Override this to return a representation of the operator state.
